tests/wpt/tests/css/css-text-decor/parsing/text-emphasis-position-valid.html

<!DOCTYPE html>
<meta charset="utf-8">
<title>CSS Text Emphasis Test: Parsing text-emphasis-position with valid values</title>
<link rel="help" href="https://drafts.csswg.org/css-text-decor/#propdef-text-emphasis-position">
<meta name="assert" content="text-emphasis-position supports the full grammar 'auto | ([over | under] || [left | right]?)'.">
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script src="/css/support/parsing-testcommon.js"></script>
<script>
test_valid_value("text-emphasis-position", "auto");
test_valid_value("text-emphasis-position", "over left");
test_valid_value("text-emphasis-position", "right under", "under");
test_valid_value("text-emphasis-position", "over");
test_valid_value("text-emphasis-position", "under");
</script>
